make multiple point lights work
PointLight class needs to know its number when passed to Shader.
This ruined the polymorphic intent, so I no longer treat
different lights through the base class pointer.
Instead, LightManager holds different light types explicitly.

+++ include/RendGL/Shader.h
@@ -4,7 +4,7 @@
 
 #include <GL/glew.h>
 
-const int MAX_POINT_LIGHTS = 5;
+const size_t MAX_POINT_LIGHTS = 5;
 
 struct UniformVariables {
     GLuint cameraPosition{ 0 };
